cv_extract/requirements.txt
docling
google-generativeai
pydantic 
python-dotenv
ollama
fastapi 
uvicorn 
python-multipart